requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.13.2
aiosignal==1.4.0
annotated-types==0.7.0
anyio==4.12.0
appopener==1.7
asyncio==4.0.0
attrs==25.4.0
audioop-lts==0.2.2
cachetools==6.2.4
certifi==2025.11.12
charset-normalizer==3.4.4
colorama==0.4.6
distro==1.9.0
edge-tts==7.2.7
frozenlist==1.8.0
google-ai-generativelanguage==0.6.15
google-api-core==2.28.1
google-api-python-client==2.187.0
google-auth==2.45.0
google-auth-httplib2==0.3.0
google-genai==1.56.0
googleapis-common-protos==1.72.0
grpcio==1.76.0
grpcio-status==1.71.2
h11==0.16.0
httpcore==1.0.9
httplib2==0.31.0
httpx==0.28.1
idna==3.11
MouseInfo==0.1.3
multidict==6.7.0
propcache==0.4.1
proto-plus==1.27.0
protobuf==5.29.5
psutil==7.2.1
pyasn1==0.6.1
pyasn1_modules==0.4.2
PyAudio==0.2.14
PyAutoGUI==0.9.54
pydantic==2.12.5
pydantic_core==2.41.5
pygame==2.6.1
PyGetWindow==0.0.9
PyMsgBox==2.0.1
pyparsing==3.3.1
pyperclip==1.11.0
PyRect==0.2.0
PyScreeze==1.0.1
python-dotenv==1.2.1
pytweening==1.2.0
pywin32==311
regex==2025.11.3
requests==2.32.5
rsa==4.9.1
sniffio==1.3.1
SpeechRecognition==3.14.5
standard-aifc==3.13.0
standard-chunk==3.13.0
tabulate==0.9.0
tenacity==9.1.2
tqdm==4.67.1
typing-inspection==0.4.2
typing_extensions==4.15.0
uritemplate==4.2.0
urllib3==2.6.2
websockets==15.0.1
wheel==0.45.1
yarl==1.22.0
